+++ internal/server/v1/observations/var_name_visitor.go
@@ -1,0 +1,46 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package observations contain code for observations.
+package observations
+
+import (
+	"go/ast"
+	"reflect"
+)
+
+// This implements ast.Visitor, used by ast.Walk, which iterates AST with a depth-first order.
+// For each ast.BinaryExpr, it extracts the "Name" property of X and Y.
+// For example, for an expression "Count_Person - Count_Person_Female[mm=USCensus]", it returns
+// {Count_Person, Count_Person_Female[mm=USCensus]}.
+type varNameVisitor struct {
+	varNameSet map[string]struct{}
+}
+
+func newVarNameVisitor() *varNameVisitor {
+	return &varNameVisitor{varNameSet: map[string]struct{}{}}
+}
+
+func (v *varNameVisitor) Visit(node ast.Node) (w ast.Visitor) {
+	switch t := node.(type) {
+	case *ast.BinaryExpr:
+		if reflect.TypeOf(t.X).String() == "*ast.Ident" {
+			v.varNameSet[t.X.(*ast.Ident).Name] = struct{}{}
+		}
+		if reflect.TypeOf(t.Y).String() == "*ast.Ident" {
+			v.varNameSet[t.Y.(*ast.Ident).Name] = struct{}{}
+		}
+	}
+	return v
+}
